Added Ollama connection
Added a slash sub command to ask the GM. This passes the prompt file and the log of the current scene to Ollama.
There is a task running the in the background watching for the ask_ai file. When it sees it, it will ask Ollama and pipe the response back into the game channel

Changed the channel global to be the channel ID.